requirements.txt
django
djangorestframework
psycopg2-binary
pillow
python-dotenv
djangorestframework-simplejwt
django-filter
coverage
